extracted shim length contructor into class, replaced some debug stuff with netty native
