+++ src/main/java/org/eclipse/californium/elements/tcp/framing/FourByteFieldPrepender.java
@@ -1,0 +1,13 @@
+package org.eclipse.californium.elements.tcp.framing;
+
+import io.netty.handler.codec.LengthFieldPrepender;
+
+public class FourByteFieldPrepender extends LengthFieldPrepender{
+
+	private static final int PREPENDED_BYTE_FIELD_LENGTH = 4;
+
+	public FourByteFieldPrepender() {
+		super(PREPENDED_BYTE_FIELD_LENGTH);
+	}
+
+}
